reverse-image-search/requirements.txt
# For Google Cloud Services
google-cloud-vision
google-cloud-storage

# For web scraping and parsing
requests
beautifulsoup4
python-dateutil
tldextract
python-multipart

# For the web application
fastapi
uvicorn[standard]

#vertexai integration
google-cloud-aiplatform  
